requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
Django==4.2.7
django-bootstrap-v5==1.0.11
django-ninja==0.22.2
fonttools==4.43.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pydantic==1.10.13
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
seaborn==0.13.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
